backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
motor
reportlab
PyPDF2
requests
twilio
python-dotenv
joblib
xgboost
langgraph
pandas
faker
numpy
scikit-learn==1.7.2
python-multipart
pillow
opencv-python-headless
scikit-image
category-encoders
google-genai

langchain
langchain-community
langchain-classic
langchain-core
langchainhub
langchain_groq
langchain_google_genai
langchain-tavily
